Java-like enums. Kinda.
- Used a combination of #defines and #includes to generate enums and
strings in upper and lower case, so that I generate Objective-C-style
enums, the names for those enums, the strings I can expect to see in
the incoming assembly-language file, and translation functions for all
those things.   Example:  You type a macro like ENUM_ENTRY(Find).  I’ll
generate, from that:  UILOpcodeFind (as an entry in the typedefed enum
“UILOpcode”), “Find,” “find,” and translation functions among those
three values.

- Also changed some of the actual assembly-language constructs.  This
was why I started this bunch of work; it quickly showed I wasn’t going
to be able to scale my assembly-language keywords (the enum) unless I
could avoid having to retype lots of copies of the same word.